celery flow solution

+++ gym_management/services/aws_services/sqs_service.py
@@ -1,9 +1,8 @@
-import boto3
+import boto3 # type: ignore
 from gym_management.settings import LOCALSTACK_URL, AWS_SECRET_ACCESS_KEY, AWS_ACCESS_KEY_ID, AWS_REGION
 
 class SQSService:
-    def __init__(self, queue_url):
-        self.queue_url = queue_url
+    def __init__(self, queue_url=None):
         self.sqs = boto3.client(
             'sqs',
             region_name=AWS_REGION,
@@ -11,22 +10,43 @@ class SQSService:
             aws_secret_access_key=AWS_SECRET_ACCESS_KEY,
             endpoint_url=LOCALSTACK_URL
         )
+        self.queue_url = queue_url
+
+    def get_queue_url(self, queue_name):
+        try:
+            response = self.sqs.get_queue_url(QueueName=queue_name)
+            self.queue_url = response['QueueUrl']
+            print(f"Queue URL retrieved: {self.queue_url}")
+            return self.queue_url
+        except Exception as e:
+            print(f"Failed to retrieve queue URL: {e}")
+            raise
 
     def receive_messages(self, max_number_of_messages=10):
+        if not self.queue_url:
+            raise ValueError("Queue URL not set. Please set it before receiving messages.")
+
         response = self.sqs.receive_message(
             QueueUrl=self.queue_url,
             MaxNumberOfMessages=max_number_of_messages,
             WaitTimeSeconds=20
         )
-        print(f"Received messages: {response.get('Messages', [])}")  # Debug line
         return response.get('Messages', [])
 
     def delete_message(self, receipt_handle):
+        if not self.queue_url:
+            raise ValueError("Queue URL not set. Please set it before deleting messages.")
         self.sqs.delete_message(
             QueueUrl=self.queue_url,
             ReceiptHandle=receipt_handle
         )
-        print(f"Message deleted with receipt handle: {receipt_handle}")  # Debug line
+        print("Message deleted from the queue.")
 
     def purge_queue(self):
-        self.sqs.purge_queue(QueueUrl=self.queue_url)
+        try:
+            response = self.sqs.purge_queue(QueueUrl=self.queue_url)
+            print("Queue purged successfully.")
+            return response
+        except Exception as e:
+            print(f"Failed to purge queue: {e}")
+            raise
